fix/break-vuex-store-in-small-files
