+++ src/store/getters.js
@@ -1,0 +1,10 @@
+const getters = {
+  students: function (state) {
+    return state.students;
+  },
+  student: function (state) {
+    return state.student;
+  },
+};
+
+export default getters;
