+++ src/store/mutations.js
@@ -1,0 +1,11 @@
+const mutations = {
+  setStudents(state, payload) {
+    state.students = payload;
+  },
+
+  setStudent(state, payload) {
+    state.student = payload;
+  },
+};
+
+export default mutations;
